Task3/twelvedata-producer/requirements.txt
kafka-python
twelvedata
websocket
requests
pandas